nits(pkg/session): remove process gossip code to a separate file
