+++ pkg/server/handlers_healthz_test.go
@@ -14,7 +14,7 @@ import (
 func TestCreateHealthzHandler(t *testing.T) {
 	gin.SetMode(gin.TestMode)
 	router := gin.New()
-	router.GET("/healthz", handleHealthz())
+	router.GET("/healthz", healthz())
 
 	tests := []struct {
 		name        string
