+++ pkg/server/handlers_machine_info.go
@@ -11,7 +11,7 @@ import (
 
 const URLPathMachineInfo = "/machine-info"
 
-// handleMachineInfo godoc
+// machineInfo godoc
 // @Summary Get machine information
 // @Description Returns detailed information about the machine including hardware specifications
 // @ID getMachineInfo
@@ -21,7 +21,7 @@ const URLPathMachineInfo = "/machine-info"
 // @Failure 404 {object} map[string]interface{} "GPUd instance not found"
 // @Failure 500 {object} map[string]interface{} "Internal server error"
 // @Router /machine-info [get]
-func (g *globalHandler) handleMachineInfo(c *gin.Context) {
+func (g *globalHandler) machineInfo(c *gin.Context) {
 	if g.gpudInstance == nil {
 		c.JSON(http.StatusNotFound, gin.H{"code": errdefs.ErrNotFound, "message": "gpud instance not found"})
 		return
