+++ pkg/server/handlers_plugins_test.go
@@ -1,0 +1,1 @@
+package server
