+++ pkg/session/serve.go
@@ -97,22 +97,6 @@ type BootstrapResponse struct {
 	ExitCode int32  `json:"exit_code,omitempty"`
 }
 
-func (s *Session) processGossip(resp *Response) {
-	if s.createGossipRequestFunc == nil {
-		return
-	}
-
-	gossipReq, err := s.createGossipRequestFunc(s.machineID, s.nvmlInstance)
-	if err != nil {
-		log.Logger.Errorw("failed to create gossip request", "error", err)
-		resp.Error = err.Error()
-		return
-	}
-
-	resp.GossipRequest = gossipReq
-	log.Logger.Debugw("successfully set gossip request")
-}
-
 func (s *Session) serve() {
 	for body := range s.reader {
 		var payload Request
